+++ ErsatzTV.Application/Search/Queries/SearchTelevisionSeasons.cs
@@ -1,0 +1,5 @@
+using ErsatzTV.Application.MediaItems;
+
+namespace ErsatzTV.Application.Search;
+
+public record SearchTelevisionSeasons(string Query) : IRequest<List<NamedMediaItemViewModel>>;