+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Limit `HLS Direct` streams to realtime speed
 - Fix `Reset Playout` button to use worker thread instead of UI thread
   - This fixes potential UI hangs and database concurrency bugs
+- Maintain watermark alpha channel (built-in transparency) using QSV acceleration
 
 ### Changed
 - Remove duplicate items from smart collections before scheduling